requirements-dev.txt
-r requirements.txt
flake8==3.8.1
pytest-django==3.9.0
coverage==5.1
pytest-cov==2.9.0
codecov==2.1.3
model-bakery==1.1.0
django-extensions==2.2.9